+++ redux/side_effect_runner.py
@@ -8,13 +8,13 @@ import inspect
 import threading
 import weakref
 from asyncio import Handle, iscoroutine
-from collections.abc import Callable
 from typing import TYPE_CHECKING, Any, Generic, cast
 
 from redux.basic_types import Event, EventHandler, TaskCreator
 
 if TYPE_CHECKING:
     import queue
+    from collections.abc import Callable
 
 
 class SideEffectRunnerThread(threading.Thread, Generic[Event]):
@@ -50,7 +50,16 @@ class SideEffectRunnerThread(threading.Thread, Generic[Event]):
                     event_handler = event_handler_
                 parameters = 1
                 with contextlib.suppress(Exception):
-                    parameters = len(inspect.signature(event_handler).parameters)
+                    parameters = len(
+                        [
+                            param
+                            for param in inspect.signature(
+                                event_handler,
+                            ).parameters.values()
+                            if param.kind
+                            in (param.POSITIONAL_ONLY, param.POSITIONAL_OR_KEYWORD)
+                        ],
+                    )
 
                 if self.create_task:
 
@@ -60,17 +69,19 @@ class SideEffectRunnerThread(threading.Thread, Generic[Event]):
                         parameters: int,
                     ) -> None:
                         if parameters == 1:
-                            result = cast(Callable[[Event], Any], event_handler)(event)
+                            result = cast('Callable[[Event], Any]', event_handler)(
+                                event,
+                            )
                         else:
-                            result = cast(Callable[[], Any], event_handler)()
+                            result = cast('Callable[[], Any]', event_handler)()
                         if iscoroutine(result):
                             await result
 
                     self.create_task(_(event_handler, event, parameters))
                 else:  # noqa: PLR5501
                     if parameters == 1:
-                        cast(Callable[[Event], Any], event_handler)(event)
+                        cast('Callable[[Event], Any]', event_handler)(event)
                     else:
-                        cast(Callable[[], Any], event_handler)()
+                        cast('Callable[[], Any]', event_handler)()
             finally:
                 self.task_queue.task_done()