+++ tests/test_autorun.py
@@ -92,7 +92,7 @@ def test_general(store_snapshot: StoreSnapshot, store: StoreType) -> None:
 
 
 def test_ignore_attribute_error_in_selector(store: StoreType) -> None:
-    @store.autorun(lambda state: cast(Any, state).non_existing)
+    @store.autorun(lambda state: cast('Any', state).non_existing)
     def _(_: int) -> int:
         pytest.fail('This should never be called')
 
@@ -100,7 +100,7 @@ def test_ignore_attribute_error_in_selector(store: StoreType) -> None:
 def test_ignore_attribute_error_in_comparator(store: StoreType) -> None:
     @store.autorun(
         lambda state: state.value,
-        lambda state: cast(Any, state).non_existing,
+        lambda state: cast('Any', state).non_existing,
     )
     def _(_: int) -> int:
         pytest.fail('This should never be called')
@@ -143,7 +143,7 @@ def test_callability(store_snapshot: StoreSnapshot, store: StoreType) -> None:
     def check(state: StateType) -> None:
         assert render() == state.value
 
-    store.subscribe(check)
+    store._subscribe(check)  # noqa: SLF001
 
 
 def test_subscription(store_snapshot: StoreSnapshot, store: StoreType) -> None: